carter base_link in centre of axle
BenchBot made incorrect assumptions on the global pose of all objects by assuming base_link was in centre of the axle of the robot's wheels. Omniverse had base_link at the absolute base of the robot causing errors in pose. This is fixed with this commit with base_link in omniverse now matching benchbot assumptions.